Make Pools instance variables.
