--- Mjml.Net/Internal/ObjectPools.cs
@@ -1,29 +1,0 @@
-﻿using System.Text;
-using Microsoft.Extensions.ObjectPool;
-
-namespace Mjml.Net.Internal
-{
-    internal static class ObjectPools
-    {
-        public static readonly ObjectPool<StringBuilder> StringBuilder =
-            new DefaultObjectPool<StringBuilder>(new StringBuilderPooledObjectPolicy());
-
-        public static readonly ObjectPool<MjmlRenderContext> Contexts =
-            new DefaultObjectPool<MjmlRenderContext>(new MjmlRenderContextPolicy());
-
-        private sealed class MjmlRenderContextPolicy : PooledObjectPolicy<MjmlRenderContext>
-        {
-            public override MjmlRenderContext Create()
-            {
-                return new MjmlRenderContext();
-            }
-
-            public override bool Return(MjmlRenderContext obj)
-            {
-                obj.Clear();
-
-                return true;
-            }
-        }
-    }
-}
